+++ app/static/lang/lang.es.js
@@ -174,6 +174,7 @@ export const lang = {
   decreaseDurText: { text: 'Disminuir duración de nota' },
   toggleDotsText: { text: 'Alternar el punteado' },
   cleanAccidText: { text: 'Verificar @accid.ges' },
+  meterConformanceText: { text: 'Verificar @metcon' },
   renumberMeasuresTestText: { text: 'Renumerar compases (prueba)' },
   renumberMeasuresExecText: { text: 'Renumerar compases (ejecutar)' },
   addIdsText: { text: 'Añadir IDs a MEI' },
@@ -1110,8 +1111,11 @@ export const lang = {
   renumberMeasureMeasuresRenumbered: { text: 'compases renumerados' },
 
   // Code checker panel (accid.ges)
-  codeCheckerTitle: {
+  accidGesCodeCheckerTitle: {
     text: 'Verificar los atributos @accid.ges (en relación con la armadura de clave, los accid. por compás y las ligaduras).',
+  },
+  metConCodeCheckerTitle: {
+    text: 'Verificando conformidad métrica (al menos una capa por pentagrama tiene número de pulso según la firma de compás).',
   },
   codeCheckerFix: { text: 'Corregir' },
   codeCheckerFixAll: { text: 'Corregir todo' },
@@ -1119,10 +1123,13 @@ export const lang = {
   codeCheckerIgnoreAll: { text: 'Ignorar todo' },
   codeCheckerCheckingCode: { text: 'Verificando código...' },
   codeCheckerNoAccidMessagesFound: { text: 'Todos los atributos accid.ges parecen correctos.' },
+  codeCheckerMeterConformanceMessage: { text: 'Todos los compases se ajustan a sus firmas métricas.' },
   codeCheckerMeasure: { text: 'Compás' },
   codeCheckerNote: { text: 'Nota' },
   codeCheckerHasBoth: { text: 'tiene ambos' },
   codeCheckerAnd: { text: 'y' },
+  codeCheckerHasADurationOf: { text: 'tiene una duración de' },
+  codeCheckerInsteadOf: { text: 'en lugar de' },
   codeCheckerRemove: { text: 'Eliminar' },
   codeCheckerFixTo: { text: 'Corregir a' },
   codeCheckerAdd: { text: 'Agregar' },